FEAT: Move incomingHeaderProcessor package from sovereign to node
