+++ cmd/sovereignnode/runType/argRunTypeComponentsCreator.go
@@ -5,9 +5,8 @@ import (
 
 	"github.com/multiversx/mx-chain-go/config"
 	"github.com/multiversx/mx-chain-go/factory/runType"
+	"github.com/multiversx/mx-chain-go/process/block/sovereign/incomingHeader"
 	"github.com/multiversx/mx-chain-go/sovereignnode/dataCodec"
-	"github.com/multiversx/mx-chain-go/sovereignnode/incomingHeader"
-
 	"github.com/multiversx/mx-sdk-abi-go/abi"
 )
